eddited admin

+++ app/Http/routes.php
@@ -21,7 +21,7 @@
 
     Route::get('/home', 'HomeController@index');
     Route::resource('user', 'UserController');
-    Route::delete('/user/{id}', 'UserController@adminDestroyer');
+    Route::delete('/user/{id}/admin', 'UserController@adminDestroyer');
     Route::get('/user/{id}/passwordchange', 'UserController@editPassword');
     Route::put('/user/{id}/passwordchange', 'UserController@storePassword');
     Route::get('/user', 'UserController@index')->middleware('admin');
